add cold reload
