+++ .gitignore
@@ -6,3 +6,4 @@ wwwroot/dist/
 .vs/
 mvc_jquery_webpack_template.csproj.user
 /Views/Shared/_Layout.cshtml
+/node_modules
