--- ClientApp/.babelrc
@@ -1,11 +1,0 @@
-﻿{
-  "presets": [
-    [
-      "@babel/preset-env",
-      {
-        "modules": false
-      }
-    ]
-  ],
-  "plugins": ["@babel/plugin-syntax-dynamic-import"]
-}
