+++ ClientApp/Views/Shared/_Layout.cshtml
@@ -1,0 +1,47 @@
+Html Webpack Plugin:
+<pre>
+  Error: Child compilation failed:
+  Entry module not found: Error: Can't resolve 'D:\Github\mvc_jquery_webpack_tem  plate\ClientApp\Views\Shared\_LayoutTemplate.cshtml' in 'D:\Github\mvc_jquery_  webpack_template\ClientApp':
+  Error: Can't resolve 'D:\Github\mvc_jquery_webpack_template\ClientApp\Views\Sh  ared\_LayoutTemplate.cshtml' in 'D:\Github\mvc_jquery_webpack_template\ClientA  pp'
+  
+  - compiler.js:79 
+    [ClientApp]/[html-webpack-plugin]/lib/compiler.js:79:16
+  
+  - Compiler.js:343 
+    [ClientApp]/[webpack]/lib/Compiler.js:343:11
+  
+  - Compiler.js:681 
+    [ClientApp]/[webpack]/lib/Compiler.js:681:15
+  
+  
+  - Hook.js:154 AsyncSeriesHook.lazyCompileHook
+    [ClientApp]/[tapable]/lib/Hook.js:154:20
+  
+  - Compiler.js:678 
+    [ClientApp]/[webpack]/lib/Compiler.js:678:31
+  
+  
+  - Hook.js:154 AsyncSeriesHook.lazyCompileHook
+    [ClientApp]/[tapable]/lib/Hook.js:154:20
+  
+  - Compilation.js:1423 
+    [ClientApp]/[webpack]/lib/Compilation.js:1423:35
+  
+  
+  - Hook.js:154 AsyncSeriesHook.lazyCompileHook
+    [ClientApp]/[tapable]/lib/Hook.js:154:20
+  
+  - Compilation.js:1414 
+    [ClientApp]/[webpack]/lib/Compilation.js:1414:32
+  
+  
+  - index.js:321 
+    [ClientApp]/[terser-webpack-plugin]/dist/index.js:321:9
+  
+  - TaskRunner.js:48 TaskRunner.run
+    [ClientApp]/[terser-webpack-plugin]/dist/TaskRunner.js:48:7
+  
+  - index.js:227 TerserPlugin.optimizeFn
+    [ClientApp]/[terser-webpack-plugin]/dist/index.js:227:18
+  
+</pre>
