--- ClientApp/package.json
@@ -1,45 +1,0 @@
-{
-  "name": "clientapp",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "webpack --mode development --progress",
-    "prod": "rimraf ../wwwroot/dist && webpack --mode production --progress"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "@babel/core": "^7.7.2",
-    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
-    "@babel/preset-env": "^7.7.1",
-    "@fortawesome/fontawesome-free": "^5.11.2",
-    "aspnet-webpack": "^3.0.0",
-    "babel-loader": "^8.0.6",
-    "babel-plugin-dynamic-import-webpack": "^1.1.0",
-    "compression-webpack-plugin": "^3.0.0",
-    "css-loader": "^3.2.0",
-    "file-loader": "^4.2.0",
-    "html-webpack-plugin": "^3.2.0",
-    "mini-css-extract-plugin": "^0.8.0",
-    "node-sass": "^4.13.0",
-    "rimraf": "^3.0.0",
-    "sass-loader": "^8.0.0",
-    "style-loader": "^1.0.0",
-    "url-loader": "^2.2.0",
-    "webpack": "^4.41.2",
-    "webpack-cli": "^3.3.10",
-    "webpack-dev-middleware": "^3.7.2",
-    "webpack-hot-middleware": "^2.25.0",
-    "webpack-dev-server": "3.9.0"
-  },
-  "dependencies": {
-    "@babel/polyfill": "7.7.0",
-    "bootstrap": "^4.3.1",
-    "jquery": "^3.4.1",
-    "moment": "^2.24.0",
-    "popper.js": "^1.16.0"
-  }
-}
