+++ appsettings.Development.json
@@ -5,5 +5,13 @@
       "System": "Information",
       "Microsoft": "Information"
     }
+  },
+ "WebRoot": "null",
+  "Port": 5000,
+  "LiveReload": {
+    "LiveReloadEnabled": true,
+    "ClientFileExtensions": ".cshtml,.css,.js,.htm,.html,.ts,.wc",
+    "ServerRefreshTimeout": 3000,
+    "WebSocketUrl": "/__livereload"
   }
 }
