+++ appsettings.json
@@ -4,5 +4,8 @@
       "Default": "Warning"
     }
   },
-  "AllowedHosts": "*"
+  "AllowedHosts": "*",
+  "LiveReload": {
+    "LiveReloadEnabled": false
+  }
 }
